backend/pyproject.toml
[project]
name = "table-project-backend"
version = "0.1.0"
description = "FastAPI backend scaffold for table_project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "arq>=0.28.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "gunicorn>=26.0.0",
    "jamdict>=0.1a11.post2",
    "jamdict-data>=1.5",
    "pydantic-settings>=2.14.0",
    "pyjwt[crypto]>=2.12.1",
    "python-multipart>=0.0.27",
    "redis>=5.3.1",
    "sqlalchemy[asyncio]>=2.0.49",
    "structlog>=25.5.0",
    "svix>=1.92.2",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.20.2",
    "pre-commit>=4.6.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.12",
]

[tool.ruff]
line-length = 99
target-version = "py312"
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "ASYNC"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
files = ["src", "tests"]
exclude = ["alembic/versions"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
